ci: add github workflow for publishing container images
Publish multiarch container images of the reference plugins at
ghcr.io/containerd/nri/plugins/<plugin>:<tag>

The tag is the git tag name for tagged versions and "unstable" for the
main branch.

The default-validator, network-device-injector and network-logger
plugins are skipped in the config as they are also missing from the
Makefile.